inputrc: Remove `menu-complete` bindings
Summary:
I used to have `<C-j>` and `<C-k>` bound to `menu-complete` and its
backward variant. Then I found that I had to switch `<C-j>` to `<C-l>`
because of bugs in GDB (reported, not fixed, not even acknowledged). But
then I discovered that there are builtins to do the same thing---`<C-n>`
and `<C-p>` are exactly what I want! So these can go.

Test Plan:
```
$ mkdir test
$ cd test
$ touch alpha avocado bravo
$ stat  # now press <C-n> to cycle through alpha, avocado, bravo
$ stat  # now press <C-p> to cycle through bravo, avocado, alpha
$ rm alpha avocado bravo
$ cd ..
$ rmdir test
```