vimrc: Enable shell-escape from vim-latexsuite
Summary:
Useful for TikZ's externalize library. Also means that I have to
exercise a greater degree of caution when running TeX code, because it
now allows arbitrary code execution by design. But this isn't too much
of a big deal, because even without shell-escape one should be careful
in running untrusted code.

Test Plan:
Create the following TeX file:
```tex
\documentclass{article}

\usepackage[paperwidth=15cm, paperheight=15cm, margin=1cm]{geometry}
\pagestyle{empty}

\usepackage{tikz}
\usetikzlibrary{external}
\tikzexternalize

\begin{document}

\centering
\begin{tikzpicture}[vertex/.style={inner sep=1pt, fill, circle}]
  \def\max{40}
  \def\baseradius{\textwidth * 0.4}
  \foreach \i in {1,...,\max} {
    \pgfmathsetmacro\theta{(\i - 1) * 360 / \max}
    \node [vertex] (\i) at (\theta : \baseradius) {};
  }
  \foreach \k in {0,...,7} {
    \pgfmathsetmacro\multiplier{\k / 8 * 0.1 + 1}
    \pgfmathsetmacro\alpha{pow((\k + 1) / 8, 2)}
    \draw [opacity=\alpha] (0, 0) circle [radius=\baseradius * \multiplier];
  }
  \foreach \i in {2,...,\max} {
    \pgfmathtruncatemacro\upto{\i - 1}
    \foreach \j in {1,...,\upto} {
      \pgfmathsetmacro\alpha{pow(1 - abs(20 - abs(\j - \i)) / 20, 4)}
      \draw [opacity=\alpha] (\i) -- (\j);
    }
  }
\end{tikzpicture}

\end{document}
```
Then, open it in vim and hit `<Leader>ll`. It should take a second or so
to compile. Compile it again and it should be instant. Also, the picture
should still be there, and should look pretty!