vimrc: Standardize clipboard bindings
Summary:
The intent is that `<Leader>c` should now essentially mean "use the
clipboard for the next action." This supersedes the old `<Leader>yd` and
`<Leader>yp` commands, whose mnemonics were pretty hackish. All
functionality is retained, and I threw in the `Y` variant because I've
found myself using that as well.

Test Plan:
Run all the commands and see that they do the right things! :-)