vimrc: Remove "space-after-bang" command line map
Summary:
I used to map `!` to `! ` on the ex command line, because I would often
run commands like `! ls dir/`, and the tab completion was helpful. But
now I'm All In on Neovim™, and so just spin up a real terminal to do
stuff like that. This mapping was definitely always controversial, so
I'm glad to see it go!

Test Plan:
Check that nvim is installed.