vimrc: Fix word-boundary matching on \part
Summary:
I'm treating `\part` as an `\item`-like command, because of its use in
the `exam` document class, and because I never write documents big
enough to actually have multiple Parts! However, this had a
long-standing bug that was driving me crazy: math-mode expressions like

    \frac{\partial f}{\partial x}

would cause the following line to be indented, because the `\part` was
being matched. I eventually realized that this was all my fault, and
that fixing the regex would fix everything. That said, the problem still
exists with the built-in commands: using `\itempo` or something causes
the same problem. One fix at a time.

Test Plan:
Write `\partial f` in a document, then hit `<CR>` and see that the newly
inserted line is at the same indentation as the previous line instead of
the next indentation level.