Initial commit of new install scripts
Summary:
This supersedes my old install process, which was far too complicated.
The new scripts are sh-compatible and trivial to read. I hope to
actually be willing to use them.

Test Plan:
I ran `install.sh` and saw that it was idempotent on my already-set-up
system. I'll try it in a virtual machine soon.